refactor: modularize middleware
