+++ app.js
@@ -13,6 +13,8 @@ if (process.env.NODE_ENV !== 'production') {
 
 const passport = require('./config/passport');
 const helpers = require('./_helpers')
+const { getReqUserFromToken } = require('./middlewares/auth')
+const { replaceReqUser } = require('./middlewares/mocha')
 
 app.use(bodyParser.urlencoded({ extended: true }))
 app.use(bodyParser.json())
@@ -26,30 +28,10 @@ app.use(methodOverride('_method'))
 app.use(passport.initialize())
 app.use(passport.session())
 
-app.use((req, res, next) => {
-  if (!req.header('Authorization')) {
-    return next()
-  }
-
-  const token = req.header('Authorization').replace('Bearer ', '')
-  jwt.verify(token, process.env.JWT_SECRET, (err, decoded) => {
-    if (err) {
-      return res.status(400).json({
-        status: 'error',
-        message: 'invalid token'
-      })
-    } else {
-      req.user = decoded
-      return next()
-    }
-  })
-})
+app.use(getReqUserFromToken)
 
 // for mocha test's requirement
-app.use((req, res, next) => {
-  req.user = helpers.getUser(req)
-  next()
-})
+app.use(replaceReqUser)
 
 app.use(routes)
