+++ middlewares/mocha.js
@@ -1,0 +1,6 @@
+module.exports = {
+  replaceReqUser: (req, res, next) => {
+    req.user = helpers.getUser(req)
+    next()
+  }
+}
